Major changes added, splash screen, music and Adrian's count down timer added ...
