+++ TileBasedPlayer20172018/Sprites/TileRef.cs
@@ -1,0 +1,16 @@
+﻿namespace Tiling
+{
+    public  class TileRef
+    {
+        public int _sheetPosX;
+        public int _sheetPosY;
+        public int _tileMapValue;
+
+        public TileRef(int x, int y, int val)
+        {
+            _sheetPosX = x;
+            _sheetPosY = y;
+            _tileMapValue = val;
+        }
+    }
+}
